+++ lib/st_button/st_button_mot.dart
@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
 
-import 'package:saturn/st_buttons/st_button_type_size_state.dart';
+import 'package:saturn/st_button/st_buttons.dart';
 import 'package:saturn/st_color_utils.dart';
-import 'package:saturn/st_buttons/st_button_enum.dart';
 
-class STButtonMOT extends STButtonTypeSizeState {
+class STButtonMOT extends StatelessWidget {
   final Widget icon;
   final String label;
   final TextStyle style;
@@ -14,6 +13,9 @@ class STButtonMOT extends STButtonTypeSizeState {
   final double radius;
   final Color borderColor;
   final double borderWidth;
+  final STButtonType buttonType;
+  final STButtonSize buttonSize;
+  final STButtonState buttonState;
 
   const STButtonMOT(
       {Key key,
@@ -21,20 +23,16 @@ class STButtonMOT extends STButtonTypeSizeState {
       @required this.label,
       this.style,
       this.onPressed,
-      @required STButtonSize buttonSize,
-      @required STButtonState buttonState,
-      STButtonType buttonType,
       this.height,
       this.width,
       this.radius,
       this.borderColor,
-      this.borderWidth})
+      this.borderWidth,
+      @required this.buttonType,
+      @required this.buttonSize,
+      @required this.buttonState})
       : assert(label != null),
-        super(
-            key: key,
-            buttonSize: buttonSize,
-            buttonState: buttonState,
-            buttonType: buttonType);
+        super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -91,5 +89,56 @@ class STButtonMOT extends STButtonTypeSizeState {
         ),
       ),
     );
+  }
+
+  Color get backgroundColor {
+    if (buttonState == STButtonState.success) {
+      return assistGreen;
+    }
+    if (buttonState == STButtonState.danger) {
+      return assistRed;
+    }
+    if (buttonState == STButtonState.focus) {
+      return secRankBlue;
+    }
+    if (buttonState == STButtonState.disable) {
+      return fourRankBlue;
+    }
+    return firRankBlue;
+  }
+
+  bool get disable {
+    if (buttonState == STButtonState.disable || onPressed == null) {
+      return true;
+    }
+    return false;
+  }
+
+  double get defaultHeight {
+    if (buttonSize == STButtonSize.small) {
+      return STButtonConst.smallHeight;
+    }
+    return STButtonConst.bigHeight;
+  }
+
+  double get defaultWidth {
+    if (buttonSize == STButtonSize.small) {
+      return STButtonConst.smallWidth;
+    }
+    return STButtonConst.bigWidth;
+  }
+
+  double get space {
+    if (buttonSize == STButtonSize.small) {
+      return STButtonConst.smallSpace;
+    }
+    return STButtonConst.bigSpace;
+  }
+
+  EdgeInsets get edgeInsets {
+    if (buttonSize == STButtonSize.small) {
+      return STButtonConst.smallEdgetInset;
+    }
+    return STButtonConst.bigEdgetInset;
   }
 }
